==> code/response_variables/01_delineating_focal_areas.Rmd <==
---
title: "Delineating focal areas"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goal**

Use GPS coordinates from resource surveys to create an outline of the focal area (sampling area). The focal area reflects where we surveyed for resources and, in treatment replicates, removed *Heliconia*. We also captured birds and set up cameras within the focal areas.

The focal area outlines will be used in two main ways:

1. To calculate resource density (i.e., resources per hectare)

2. In the analysis of telemetry data, to determine the amount of time radio-tagged birds spent in the area

**Approach**

Calculate minimum convex polygons (MCP) for the GPS coordinates.

```{r setup, include = FALSE}
library(sp)
library(sf)
library(adehabitatHR)
library(tidyr)
library(dplyr) #Load after spatial packages so that select() does not get superseded
library(purrr)
```

## Step 1: Import data

```{r}
#1. Subset to one calorie estimate. Dataset has a high and low calorie estimate, but only need one here
#2. Remove rows with NA coordinates
#3. Create a factor level combining patch and year
resource_data <- read.csv("../../data/import/data/resource_counts/Resource_counts_2016-2018.csv") %>%
  filter(!is.na(x_utm) & !is.na(y_utm)) %>%
  mutate(year_patch = as.factor(paste(year, patch, sep = "_")))
```

## Step 2: Create minimum convex polygons per patch and year

Note: there are important syntax differences between the mcp() function in package **adehabitat** vs. package **adehabitatHR**.

**adehabitat** does not take a SpatialPointsDataFrame and takes an "id" argument (in addition to the "xy" argument for the coordinates)
**adehabitatHR** takes a SpatialPointsDataFrame and ONLY the xy argument, but instead of indicating coordinates this column indicates the id!

I will use adehabitatHR here.

Making SpatialPointsDataframe 
```{r}
#To calculate MCP, just want the coordinates and IDs (ie sites)

#Coordinates
coords <- resource_data %>%
  dplyr::select(x_utm, y_utm)

coords_utm <- SpatialPoints(coords, proj4string = CRS("+proj=utm +zone=17 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"))
coords_ll <- spTransform(coords_utm, CRS("+proj=longlat")) #Coords in lat/long, need this to export as kml

#IDs
ids <- resource_data %>%
  dplyr::select(year_patch)

#Combining to make SpatialPointsDataframe
spdf_utm <- SpatialPointsDataFrame(coords = coords_utm, data = ids)
spdf_ll <- SpatialPointsDataFrame(coords = coords_ll, data=ids) #Coords in lat/long, for export as kml

View(spdf_utm@data)
View(spdf_utm@coords)
```

Creating 100% minimum convex polygons for all the sites and years
```{r}
#Polygons with UTM
#In xy argument, indicate just the column with the 'animal ID' (here, 'year_patch')
focal_areas_utm <- adehabitatHR::mcp(xy = spdf_utm[,1], percent = 100, unout = "ha")

plot(focal_areas_utm)
class(focal_areas_utm) #This is a Spatial Polygons Data Frame.
View(focal_areas_utm@data) #This gives focal areas in hectares (but also see below)

#Polygons with lat/long
focal_areas_ll <- adehabitatHR::mcp(xy = spdf_ll[,1], percent = 100, unout = "ha")
```

## Step 3: Calculate area (in hectares) of minimum convex polygons
```{r}
#Convert to sf, which will give focal areas in hectares
areas <- st_as_sf(focal_areas_utm) %>%
  separate(id, into = c("year", "patch"), remove = FALSE) %>%
  mutate(mcp_percent = 100)
```

Problem: in 2016, we only surveyed for resources in the treatment sites, but we still need focal area outlines for the 2016 controls

Solution: substitute 2017 patch areas/outlines for 2016 controls that are missing resource counts (and therefore focal area outlines). In 2017, we used the same sites as 2016.
```{r}
#Get experiment info
exp_info <- read.csv("../../data/import/data/Experiment_dates_2016-2018.csv") %>%
  distinct(year, patch, control_treatment) %>%
  mutate_all(as.character)

#2016 controls are missing focal area outlines
missing_replicates <- exp_info %>%
  filter(year == 2016 & control_treatment == "control")

#Need to substitute outlines from 2017
substitute_replicates <- areas %>%
  filter(patch %in% missing_replicates$patch) %>%
  filter(year == 2017) %>%
  mutate(year = "2016")

#Complete dataset
areas <- areas %>%
  bind_rows(substitute_replicates) %>%
  arrange(mcp_percent, year, patch) %>%
  left_join(exp_info)
```

Add centroids of each patch and adjust area of site that contained lake.
```{r}
areas_and_centroids <- areas %>%
  mutate(centroid = st_centroid(geometry)) %>%  #class of this column is "sfc_POINT" "sfc" 
  mutate(centroid_list = map(centroid, st_coordinates), #this is a list instead of a sf object
         x = unlist(map(centroid_list, 1)),
         y = unlist(map(centroid_list, 2))) %>%
  as.data.frame() %>%
  dplyr::select(-geometry, -centroid, -centroid_list) %>%
  mutate(area = ifelse(patch == 201, area - 0.8, area)) #Subtract 0.8ha from patch 201 (Cantaros) - this was grass/lake
```

Export focal areas (in hectares)
```{r}
write.csv(areas_and_centroids , "../../data/export/intermediate/Focal_area_sizes_and_centroids.csv") 
```

Export focal area outlines as ESRI shapefiles and as KML files (for Google Earth)
```{r, eval = FALSE}
#'areas' is still an sf dataframe with geometry
areas_for_export <- areas %>%
  dplyr::select(-area)

#Export as ESRI shapefile
st_write(areas_for_export, "../../data/export/intermediate/focal_areas_2016-2018.shp", driver = "ESRI Shapefile") 

#Export as KML file. Rename 'id' so it shows up in the Google Earth 'description' field
st_write(rename(areas_for_export, Description = id), "../../data/export/intermediate/focal_areas_2016-2018.kml", driver = "kml", delete_dsn = TRUE) 
```
