==> code/response_variables/10_processing_summarizing_pollen_tubes.Rmd <==
---
title: "Processing and summarizing pollen tube data"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Context and goals**

These are the data for pollination success; I determined whether or not flowers were pollinated based on the presence of pollen tubes in the style. This was done by examining the style under a microscope. While looking for pollen tubes, the observer did not know what site or time period the styles came from; they just knew the species.

In a previous step (in the data repo), I added in the data about which site/date the styles (flowers) came from. This was recorded on the vial that held the styles, before lab work began. The date written on the vial was the date that styles were *collected*. This is *not* necessarily the date the flowers were open/could be pollinated. Here, I assume that was the day before styles were collected, which is a reasonable assumption for plant species with single-day flowers (such as *Heliconia*).

In this script, I:

- Explore the data and determine which plant species had enough pollen tubes for analysis (pollen tubes were pretty rare overall)

- Summarize data to create my response variable: proportion of flowers that were pollinated.

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)

library(dplyr)
library(tidyr)
library(ggplot2)
```

## Step 1: Import data
```{r step1}
tubes <- read.csv("../../data/import/data/pollen_tubes/Pollen_tubes_2016-2018.csv")

plant_names <- read.csv("../../data/import/data/Plant_codes_species_list.csv") %>%
  select(plant_species, plant_scientific_name) %>%
  filter(!is.na(plant_species) & plant_species != "") #Get rid of blank rows
```

## Step 2: Identify styles that should be included in analysis

Not all dates and plants should be analyzed. 

These data should be **included**:

- Numbered experimental days (i.e. days assigned pre/post), which are the days between the capture 1 and capture 2. These days also correspond to when we have telemetry and camera data.

- 'Focal' plants (plants at stations)
  
- Plants with styles (vials) from pre *and* post periods, because I'm interested in the pre-to-post change

These data should be **excluded**:

- Cover day in treatment sites
  
- Day after covering in treatment sites ('adjustment day')
  
- Instances were flowers collected in the 'post' period were known to be open during the 'pre' period (i.e., blurred delineation between experimental periods)
  
- 'Non-focal' plants; these styles were collected haphazardly from around the focal area, and are probably especially noisy. Also, I cannot be sure if same plant was sampled pre and post.

```{r step4a}
#Note: initially I ran these filtering steps on just the vial/experiment info to identify 'priority' vs. 'non-priority' vials for microscope work. But now that the lab work is done, I'm filtering the actual data. For instance, I ended up analyzing quite a few 'test' (non-priority) styles while ironing out methods, and these are still included in the pollen tube dataset. I think I also analyzed non-focal plants initially.
tubes02 <- tubes %>%
  filter(!(exp_phase == "NA" | is.na(exp_phase))) %>%
  filter(focal_nf == "focal") %>%
  filter(cover_day != "1") %>% #Only applies to treatment sites
  filter(day_after_cover_day != "1") %>% #Only applies to treatment sites
  filter(!(grepl("open during pre", vial_notes, ignore.case = TRUE) & exp_phase == "post"))
  
#Plants that have data pre and post
plants_to_keep <- tubes02 %>%
  group_by(year, patch, plant_species, exp_phase, focal_nf) %>%
  summarise(num_vials = n()) %>%
  pivot_wider(names_from = exp_phase, values_from = num_vials) %>%
  select(year:focal_nf, pre, post) %>%
  filter(pre > 0 & post > 0) %>%
  select(-pre, -post) %>%
  mutate(keep = 1)

tubes03 <- tubes02 %>%
  left_join(plants_to_keep) %>%
  filter(keep == 1) %>%
  select(-keep)
```

Also remove data that seemed suspect during microscope work

- Data flags currently include situations where I noted that the style was (a) potentially from a different species than the one I thought I was analyzing, (b) a partial style, so there could have been a pollen tube in the missing part, and (c) a test style not meant for analysis.

```{r step4b}
unique(tubes03$flag)

tubes04 <- tubes03 %>%
  filter(is.na(flag) | flag == "NA") %>%
  filter(!is.na(tube_presence))

tubes_for_analysis <- tubes04 %>%
  mutate(year_patch = paste(year, patch, sep = "_")) %>%
  select(-focal_nf)
```

## Step : Explore data, part I

How many styles had pollen tubes for each species?

*Note: I focused on analyzing HETO, HAPA, PAV-ROJ-AMA, because I felt most confident about my pollen tube IDs for those species (and because RABO had no pollen tubes).*
```{r step5a, fig.height = 8, fig.width = 12}
presence_plot <- tubes_for_analysis %>%
  mutate(tube_presence = as.factor(tube_presence)) %>%
  ggplot(data = ., aes(x = tube_presence, fill = tube_presence)) +
    geom_histogram(stat = "count") + 
    facet_grid(. ~ plant_species) + 
    labs(x = "Tube presence (0/1)", y = "Number of styles with at least 1 pollen tube", fill = "Tube presence") +
    scale_fill_manual(values = c("red2", "cornflowerblue")) +
    theme_bw(base_size = 18)

presence_plot
```

How many replicates and years represented for each species?
```{r step5b, fig.height = 12, fig.width = 8}
replicate_style_plot <- tubes_for_analysis %>%
  ggplot(data = ., aes(x = year_patch, fill = factor(year), alpha = control_treatment)) +
    geom_histogram(stat = "count", size = 1.25) + 
    facet_grid(plant_species~.) + 
    labs(x = "Replicate", y = "Number of styles analyzed") +
    theme_bw(base_size = 18) +
    scale_fill_manual(values = c("cornflowerblue", "orange", "red2")) +
    scale_alpha_manual(values = c(0.5, 1)) +
    guides(color = guide_legend(override.ae = list(fill = NA))) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
    theme(legend.title = element_blank())

replicate_style_plot
```

## Step 6: Summarize to level of experimental period (pre/post)

```{r step6}
#Proportion of styles with at least one tube - metric used by Hadley et al. 2014, Woods et al. 2022, and I think Kormann et al. as well. Essentially the % of flowers that were pollinated
proportion_pollinated_sum <- tubes_for_analysis %>%
  group_by(plant_species, year, patch, control_treatment, exp_phase, tube_presence) %>%
  summarise(styles = n()) %>%
  ungroup() %>%
  pivot_wider(names_from = tube_presence, values_from = styles, values_fill = 0) %>%
  rename(styles_with_tubes = "1", styles_no_tubes = "0") %>%
  mutate(styles = styles_with_tubes + styles_no_tubes) %>%
  mutate(prop_with_tubes = styles_with_tubes/styles) %>%
  filter(plant_species == "HAPA" | plant_species == "HETO" | plant_species == "PAV-ROJ-AMA" | plant_species == "RABO") %>%
  left_join(plant_names)

#How many styles ended up in the final sample?
styles_analyzed <- proportion_pollinated_sum %>%
  group_by(plant_species) %>%
  summarise(styles = sum(styles))

styles_analyzed

#Check for complete separation. Are there data for each experimental period, control AND treatment?
styles_analyzed_per_category <- proportion_pollinated_sum %>%
  group_by(plant_species, control_treatment, exp_phase) %>%
  summarise(styles_with_tubes = sum(styles_with_tubes),
            styles_no_tubes = sum(styles_no_tubes))

styles_analyzed_per_category

#Minimum number of total tubes/style. Did not use this method because past researchers in our lab have used the other metric.
# tubes_per_style_sum <- tubes_for_analysis %>%
#   group_by(plant_species, year, patch, control_treatment, exp_phase, focal_nf) %>%
#   summarise(tubes = sum(tubes_total_min, na.rm = TRUE),
#             styles = n()) %>%
#   ungroup() %>%
#   mutate(tubes_per_style = tubes/styles)
```

## Step 7: Explore data, part II

Compare pre/post, control/treatment
```{r step7, fig.height = 5, fig.width = 5}
#Functions to help with plotting
give_n <- function(x){
  
   return(c(y = max(x) + 0.5, label = length(x))) #Multiplier adjusts position
  
}

make_boxplot <- function(dataset, plant_species, yvar, ylabel){
    
    plot <- ggplot(data = dataset, aes(x = control_treatment, y = .data[[yvar]], fill = exp_phase)) +
      geom_boxplot(position = position_dodge(0.85), outlier.shape = NA) +
      geom_point(aes(group = exp_phase), alpha = 0.5, position = position_jitterdodge(dodge.width = 0.85), size = 2) +
      theme_bw(base_size = 18) +
      labs(x = "", y = ylabel, fill = "") +
      ggtitle(plant_species) +
      scale_fill_manual(values = c("cornflowerblue", "red2"))

    return(plot)
    
}

#Proportion of styles with at least one pollen tube
proportion_pollinated_plots <- proportion_pollinated_sum %>% 
  select(plant_species) %>%
  distinct() %>%
  mutate(data = purrr:::map(plant_species, ~filter(proportion_pollinated_sum, plant_species == .x))) %>%
  mutate(plot = purrr:::map2(data, plant_species, ~make_boxplot(dataset = .x, plant_species = .y, yvar = "prop_with_tubes", ylabel = "% flowers pollinated")))

proportion_pollinated_plots$plot

#There are no PAV-ROJ-AMA flowers pollinated in the control 'pre'...this will lead to complete separation in the model and prevent further analysis.
```

## Step 8: Export data
```{r step8}
#Filtered data (unsummarized)
write.csv(tubes_for_analysis, "../../data/export/intermediate/Pollen_tube_data_filtered_for_analysis.csv", row.names = FALSE)

#Summarized data for analysis
write.csv(proportion_pollinated_sum, "../../data/export/for_analysis/Pollen_tube_pollination_success.csv", row.names = FALSE)
```
