==> code/response_variables/12_summarizing_camera_methods.Rmd <==
---
title: "Summarizing camera methods"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)

library(tidyr)
library(dplyr)
library(ggplot2)
library(sf)
library(purrr)
```

**Overall goals:**

1. Summarize video effort in various dimensions:

- Number of cameras per replicate
- Number of plant species per replicate

[Note: video hours have already been summarizing in script summarizing camera data]

2. Calculate additional summary stats and information for methods/supplemental methods:

- Percentage of cameras at stations ('station' = cluster of plants + cameras, 2 stations per site)
- Distance between stations
- What plant species were in pots (we brought these species in to create a 'floral array' at each station)

## Step 1: Get data
```{r}
#Video data to analyze
data_experiment <- read.csv("../../data/export/intermediate/Camera_data_filtered_for_experiment.csv")

#Experimental timeline data (what happened on each day)
exp_info <- read.csv("../../data/import/data/Experiment_dates_2016-2018.csv")

#Camera location data
camera_locations <- read.csv("../../data/import/data/cameras/Camera_locations_2016-2018.csv")

#Plant scientific name, etc. for supplement
plant_info <- read.csv("../../data/import/data/Plant_codes_species_list.csv")

#Function to calculate basic summary stats (mean, median, etc.)
source("../../code/helper_functions/Calculate_basic_summary_stats.R")
```

## Step 2: Calculate number of plant species + number of cameras per replicate 

I can either calculate these numbers from:

1. The dataset of hummingbird sightings that will be used in analysis (i.e., the one I've been working with so far)

OR 

2. The dataset of camera locations (data recorded during camera installation)

The first option is more conservative, because not all cameras that were set up yielded usable data. For example, perhaps the focal plant had no flowers or data were unavailable for both experimental periods, etc. However, in the methods section, it might be a bit more natural to say "We installed X cameras per replicate" rather than "We obtained usable data from X cameras per replicate".

**Effort summary from sighting dataset (final, usable data)**
```{r step2a, message = FALSE}
#Number of cameras per replicate (there can be multiple plant species per camera)
cameras_per_replicate = data_experiment %>%
  distinct(year, patch, control_treatment, camera_num) %>%
  group_by(year, patch, control_treatment) %>%
  summarise(num_cameras = n()) %>%
  ungroup() %>%
  filter(!is.na(patch))

cameras_per_replicate_sum = cameras_per_replicate %>%
    calculate_basic_summary_stats(data = ., variable = num_cameras)

#How much does the number of cameras/site vary between control and treatment replicates?
cameras_per_replicate %>%
  mutate(control_treatment = factor(control_treatment, levels = c("control", "treatment"), labels = c("Control", "Treatment"))) %>%
  ggplot(data = ., aes(x = control_treatment, y = num_cameras, colour = control_treatment)) +
    geom_violin(position = position_dodge(width = 1), trim = FALSE) +
    geom_boxplot(position = position_dodge(width = 1), width = 0.25) +
    geom_point(position = position_jitterdodge(dodge.width = 1, jitter.width = 0.2, jitter.height = 0.2), alpha = 0.5) +
    labs(y = "Number of cameras", x = "", colour = "") +
    scale_colour_manual(values = c("#0E0D37", "#BA0022"), guide = "none") +
    theme_bw(base_size = 18)

#Number of plant species per replicate
species_per_replicate_sum = data_experiment %>%
  distinct(year, patch, control_treatment, plant_species) %>%
  group_by(year, patch, control_treatment) %>%
  summarise(num_species = n()) %>%
  ungroup() %>%
  filter(!is.na(patch)) %>%
  calculate_basic_summary_stats(data = ., variable = num_species)
  
#Finally, just out of curiosity...which plant species are present most reliably across the different replicates?
replicates_per_species <- data_experiment %>%
  distinct(year, patch, control_treatment, plant_species) %>%
  group_by(plant_species) %>%
  summarise(num_replicates = n()) %>%
  ungroup() %>%
  arrange(desc(num_replicates))

head(replicates_per_species)
```

**Effort summary from camera setup dataset (record of cameras installed)**
```{r step2b, message = FALSE}
#Number of cameras per replicate (there can be multiple plant species per camera)
cameras_per_replicate_setup = camera_locations %>%
  distinct(year, patch, control_treatment, camera_number_video_data) %>%
  group_by(year, patch, control_treatment) %>%
  summarise(num_cameras = n()) %>%
  ungroup() %>%
  filter(!is.na(patch))

cameras_per_replicate_setup_sum <- cameras_per_replicate_setup %>%
  calculate_basic_summary_stats(data = ., variable = num_cameras)

#Number of plant species per replicate
species_per_replicate_setup_sum = camera_locations %>%
  distinct(year, patch, control_treatment, plant_species) %>%
  group_by(year, patch, control_treatment) %>%
  summarise(num_species = n()) %>%
  ungroup() %>%
  filter(!is.na(patch)) %>%
  calculate_basic_summary_stats(data = ., variable = num_species)
  
#Number of plant species that were in pots (i.e., brought in to be part of focal array). This data was only recorded in 2017-2018
potted_plants_per_replicate_setup_sum <- camera_locations %>%
  filter(potted_y_n == "Y") %>%
  distinct(year, patch, control_treatment, camera_number_video_data) %>%
  group_by(year, patch, control_treatment) %>%
  summarise(num_plants = n()) %>%
  ungroup() %>%
  calculate_basic_summary_stats(data = ., variable = num_plants)

potted_plants_per_replicate_setup_sum
```

**Compare methods**
```{r step2c, message = FALSE}
cameras_per_replicate_sum
cameras_per_replicate_setup_sum

species_per_replicate_sum
species_per_replicate_setup_sum
```

*As expected, more cameras were set up than yielded usable data. Results are similar for number of plant species as well.*

## Step 3: Calculate details of camera spatial arrangement

- Percent of cameras at stations
- Distance between stations

**Calculate percentage of cameras at stations**
```{r step3a, message = FALSE}
#Note: these data might be a bit noisy, as I'm not sure how consistent field techs were in determining whether plant was 'at station' or not (i.e., estimating how far from focal Heliconia plant it was and recorded it as being 'at station' or not)
cameras_per_replicate_station_setup = camera_locations %>%
  filter(station == "1" | station == "2") %>%
  distinct(year, patch, control_treatment, camera_number_video_data) %>%
  group_by(year, patch, control_treatment) %>%
  summarise(num_cameras_station = n()) %>%
  ungroup()

percent_cameras_at_station_setup_sum = cameras_per_replicate_setup %>%
  left_join(cameras_per_replicate_station_setup) %>%
  mutate(percent_at_station = num_cameras_station/num_cameras * 100) %>%
  calculate_basic_summary_stats(data = ., variable = percent_at_station)
  
percent_cameras_at_station_setup_sum
```

*On average, >80% of cameras at a station*

**Calculate distance between stations**

Stations comprised a variety of different flowering plant species -- some naturally occurring, some brought it in in pots -- and was centered around a 'focal' Heliconia plant. There were two stations per site. How far apart were these stations? 

Get data and write helper function for distance calculation
```{r step3b, message = FALSE, warning = FALSE}
#Projection to use
projstring <- "+proj=utm +zone=17 +ellps=WGS84 +units=m +no_defs" 

#Function to calculate distance between locations of interest
calculate_distance <- function(data, group_id){
  
  #Select data from same group
  data_same_group <- data %>%
    filter(replicate_id == group_id)
  
  #If there is >1 element, calculate distance between them
  if(length(data_same_group$replicate_id) > 1){
    
    distance <- st_distance(data_same_group$geometry, by_element = FALSE)

    #First element in vector is zero; second element is actual distance
    return(distance[[2]]) 
  
  }
  
  return(NULL)

}

#Make a empty table to be filled in using purrr::map(). Replicate ID (patch + year combination) will be the group of interest (i.e., how far apart stations are within each replicate)
base_table <- camera_locations %>%
  distinct(year, patch) %>%
  mutate(replicate_id = paste(patch, year, sep = "_"))
```

**Calculate distance using average station coordinates**
```{r step3c, message = FALSE}
#First, average coordinates assigned to each station (sometimes there are different coordinates, but presumably close to each other)
camera_stations <- camera_locations %>%
  distinct(year, patch, control_treatment, camera_number_video_data, plant_species, station, x, y) %>%
  filter(station == 1 | station == 2) %>%
  filter(!(is.na(x) | is.na(y))) %>%
  group_by(year, patch, station) %>%
  summarise(x = mean(x),
            y = mean(y)) %>%
  mutate(replicate_id = paste(patch, year, sep = "_"))

camera_stations_spatial <- st_as_sf(camera_stations, coords = c("x", "y"))
camera_stations_spatial <- st_set_crs(camera_stations_spatial, value = projstring)

station_distance = base_table %>%
  mutate(distance = map(replicate_id, ~calculate_distance(data = camera_stations_spatial, group_id = .))) %>%
  unnest(distance) %>%
  arrange(replicate_id)

#1. Remove 'unit' attributes (meters) 2. Summarize
station_distance_sum <- station_distance %>%
  mutate(distance = units::drop_units(distance)) %>%
  calculate_basic_summary_stats(data = ., variable = distance)
```

**Calculate distance using distance of focal Heliconia locations**

Stations should be centered around focal Heliconia, so results should be similar as station results
```{r step3d, message = FALSE}
focal_heto_locations <- camera_locations %>%
  distinct(year, patch, control_treatment, camera_number_video_data, plant_species, station, x, y) %>%
  filter((station == 1 | station == 2)) %>% #Note: by only looking at station numbers, this omits 2017 p10 focal HETO that was not at station
  filter(plant_species == "HETO") %>%
  filter(!(is.na(x) | is.na(y))) %>%
  group_by(year, patch, station) %>%
  summarise(x = mean(x),
            y = mean(y)) %>%
  mutate(replicate_id = paste(patch, year, sep = "_"))

focal_heto_locations_spatial <- st_as_sf(focal_heto_locations, coords = c("x", "y"))
focal_heto_locations_spatial <- st_set_crs(focal_heto_locations_spatial, value = projstring)

station_distance_focal_heto = base_table %>%
  mutate(distance = map(replicate_id, ~calculate_distance(data = focal_heto_locations_spatial, group_id = .))) %>%
  unnest(distance) %>%
  arrange(replicate_id)

#1. Remove 'unit' attributes (meters) 2. Summarize
station_distance_focal_heto_sum <- station_distance_focal_heto %>%
  mutate(distance = units::drop_units(distance)) %>%
  calculate_basic_summary_stats(data = ., variable = distance)
```

**Compare methods**
```{r step3e, message = FALSE}
station_distance_sum
station_distance_focal_heto_sum
```

*Results are indeed similar if use station locations vs. focal Heliconia. Median of ~50m, mean of ~60 m +/- 44 m SD*

## Step 4: Summarize which plant species were present in focal arrays at stations (for table in Supplemental Methods)

```{r step4, message = FALSE}
#Note: potted status was only recorded in 2017-2018, but plants used in these years encompass all of 2016 species except IMPO (which did not yield any usable camera data anyway)
floral_array_species <- camera_locations %>%
  filter(potted_y_n == "Y") %>%
  distinct(plant_species)

floral_array_species 

floral_array_species_supp <- floral_array_species %>%
  left_join(plant_info) %>%
  filter(!is.na(plant_family)) %>%
  arrange(plant_family, plant_scientific_name) %>%
  filter(plant_species != "HETO") #Remove HETO. Not really part of floral array/potted plants

#Where there any species recorded in setup data that didn't end up in usable data?
check01 <- floral_array_species_supp %>%
  filter(!(plant_species %in% data_experiment$plant_species))

#Export
write.csv(floral_array_species_supp, "../../data/export/supp_tables/List_of_floral_array_species.csv", row.names = FALSE)
```

## Step 5: Calculate when floral arrays were established (i.e., how many days in advance of experiment start)

Here, I will define the experiment start date as the first capture day (or first video day if there was no capture).

Plants were brought in ahead of this date, on the 'setup' day (or first setup day when there was more than one; the second setup day was for camera setup).

I need to calculate the number of days between floral array setup and the experiment start date.

```{r step6, warning = FALSE}
#Identify the start dates for each replicate. Start date is generally capture_1, expect in replicates without captures (2016 p204, p49). Note that experimental day numbering ('exp_day') column in dataset starts *after* capture day, so I need to update that column to actually identify the start date
start_dates <- exp_info %>%
  select(year, patch, control_treatment, exp_phase, date, exp_day) %>%
  filter(exp_phase == "capture_1" | exp_phase == "capture_2" | exp_phase == "pre" | exp_phase == "post") %>%
  mutate(date = lubridate::mdy(date)) %>%
  arrange(year, patch, date) %>%
  group_by(year, patch) %>%
  mutate(exp_day_renumber = 1:length(year)) %>%
  ungroup() %>%
  filter(exp_day_renumber == 1) %>%
  select(-exp_phase, -starts_with("exp_day")) %>%
  rename(date_start = date)
  
#Identify the setup dates
setup_dates <- exp_info %>%
  select(year, patch, control_treatment, exp_phase, date, exp_day) %>%
  filter(exp_phase == "setup" | exp_phase == "setup_1") %>%
  mutate(date = lubridate::mdy(date)) %>%
  select(-exp_phase, -exp_day) %>%
  rename(date_setup = date)

#Calculate days between setup and capture
establishment_days <- start_dates %>%
  left_join(setup_dates) %>%
  mutate(days = date_start - date_setup) %>%
  mutate(days = ifelse(is.na(days), 0, days)) %>%
  mutate(days = as.numeric(days))

#Summary stats on days between setup and capture
establishment_days_sum <- calculate_basic_summary_stats(data = establishment_days, variable = days)
establishment_days_sum
```
